requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
Office365-REST-Python-Client==2.5.3
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
pytest==7.4.2